Use index from HasIndex in IndexedMap
