+++ src/openvic-simulation/types/IndexedMap.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
 #include <concepts>
-#include <vector>
 
 #include "openvic-simulation/types/fixed_point/FixedPointMap.hpp"
+#include "openvic-simulation/types/HasIndex.hpp"
 #include "openvic-simulation/utility/ForwardableSpan.hpp"
 #include "openvic-simulation/utility/Getters.hpp"
 #include "openvic-simulation/utility/Logger.hpp"
@@ -310,10 +310,14 @@ namespace OpenVic {
 		}
 
 		constexpr size_t get_index_from_item(key_ref_type key) const {
-			if (has_keys() && keys.data() <= &key && &key <= &keys.back()) {
-				return std::distance(keys.data(), &key);
+			if constexpr (utility::is_derived_from_specialization_of<Key, HasIndex>) {
+				return key.get_index();
 			} else {
-				return 0;
+				if (has_keys() && keys.data() <= &key && &key <= &keys.back()) {
+					return std::distance(keys.data(), &key);
+				} else {
+					return 0;
+				}
 			}
 		}
